% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quasar_spec}
\alias{quasar_spec}
\title{Quasar spectrum}
\format{
A list with 9 elements:
\describe{
  \item{col}{}
  \item{hdr}{}
  \item{colNames}{}
  \item{colUnits}{}
  \item{TNULLn}{}
  \item{TSCALn}{}
  \item{TZEROn}{}
  \item{TDISPn}{}
  \item{header}{}
}
}
\source{
\url{http://sdss3.org/}
}
\usage{
quasar_spec
}
\description{
An electromagnetic spectrum measured by the Sloan Digital Sky Survey of a
quasar at redshift z ~ 2.953
}
\keyword{datasets}
