% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendfilter.R
\name{trendfilter}
\alias{trendfilter}
\title{User-friendly trend filtering (under construction. do not use.)}
\usage{
trendfilter(x, y, weights = NULL, control.list)
}
\arguments{
\item{x}{The vector of the observed inputs.}

\item{y}{The vector of the observed outputs.}

\item{weights}{A vector of weights for the observed outputs. These are defined as 
\code{weights = 1 / sigma^2}, where \code{sigma} is a vector of standard 
errors of the uncertainty in the output measurements. \code{weights} should 
either have length equal to 1 (corresponding to observations with a constant 
(scalar) variance of \code{sigma = 1/sqrt(weights)}) or length equal to 
\code{length(y)} (i.e. heteroskedastic outputs).}

\item{control.list}{\itemize{
\item{\code{modeling.control()}:\cr \code{k, lambda, nlambda, V, validation.error.type}}
\item{\code{evaluation.conrol()}:\cr \code{n.eval, x.eval}}
\item{\code{algorithm.control()}:\cr \code{thinning, max_iter, obj_tol}}
}}
}
\value{
An object of class 'trendfilter'. This is a list with the 
following elements:
\item{x.eval}{The grid of inputs the optimized trend filtering estimate was 
evaluated on.}
\item{tf.estimate}{The optimized trend filtering estimate of the signal, 
evaluated on \code{x.eval}.}
\item{validation.method}{"cv"}
\item{V}{The number of folds the data are split into for the V-fold cross
validation.}
\item{validation.error.type}{Type of error that validation was performed on. 
One of \code{c("WMAE","WMSE","MAE","MSE")}.}
\item{lambda}{Vector of hyperparameter values tested during validation. This
vector will always be returned in descending order, regardless of the 
ordering provided by the user. The indices \code{i.min} and \code{i.1se}
correspond to this descending ordering.}
\item{error}{Vector of cross validation errors for the given hyperparameter 
values.}
\item{se.error}{The standard errors of the cross validation errors.
These are particularly useful for implementing the ``1-standard-error rule''. 
The 1-SE rule favors a smoother trend filtering estimate by, instead of 
using the hyperparameter that minimizes the CV error, instead uses the 
largest hyperparameter that has a CV error within 1 standard error of the
smallest CV error.}
\item{lambda.min}{Hyperparameter value that minimizes the SURE error curve.}
\item{lambda.1se}{The largest hyperparameter value that is still within one
standard error of the minimum hyperparameter's cross validation error.}
\item{df}{Vector of effective degrees of freedom for trend filtering
estimators fit during validation.}
\item{df.min}{The effective degrees of freedom of the optimally-tuned trend 
filtering estimator.}
\item{df.1se}{The effective degrees of freedom of the 1-stand-error rule
trend filtering estimator.}
\item{i.min}{The index of \code{lambda} that minimizes the cross validation 
error.}
\item{i.1se}{The index of \code{lambda} that gives the largest hyperparameter
value that has a cross validation error within 1 standard error of the 
minimum of the cross validation error curves.}
\item{x}{The vector of the observed inputs.}
\item{y}{The vector of the observed outputs.}
\item{weights}{A vector of weights for the observed outputs. These are
defined as \code{weights = 1/sigma^2}, where \code{sigma} is a vector of 
standard errors of the uncertainty in the measured outputs.}
\item{fitted.values}{The trend filtering estimate of the signal, evaluated at
the observed inputs \code{x}.}
\item{residuals}{\code{residuals = y - fitted.values}.}
\item{k}{The degree of the trend filtering estimator.}
\item{thinning}{If \code{TRUE}, then the data are preprocessed so 
that a smaller, better conditioned data set is used for fitting. When set to
\code{NULL}, the default, function will auto detect whether thinning should 
be applied (i.e., cases in which the numerical fitting algorithm will 
struggle to converge).}
\item{max_iter}{Maximum iterations allowed for the trend filtering 
convex optimization 
(\href{http://www.stat.cmu.edu/~ryantibs/papers/fasttf.pdf}{Ramdas & Tibshirani 2015}). 
Increase this if the trend filtering estimate does not appear to have fully 
converged to a reasonable estimate of the signal.}
\item{obj_tol}{The tolerance used in the convex optimization stopping 
criterion; when the relative change in the objective function is less than 
this value, the algorithm terminates. Decrease this if the trend filtering 
estimate does not appear to have fully converged to a reasonable estimate of 
the signal.}
}
\description{
\loadmathjax{} This is a simple high-level function for
user-friendly trend filtering. This should be a very long paragraph...
}
\details{
This should be a very detailed description...
}
\references{
\enumerate{
\item{Politsch et al. (2020a). Trend filtering – I. A modern 
statistical tool for time-domain astronomy and astronomical spectroscopy. 
\emph{Monthly Notices of the Royal Astronomical Society}, 492(3), 
p. 4005-4018.
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{[Link]}} \cr

\item{Politsch et al. (2020b). Trend Filtering – II. Denoising 
astronomical signals with varying degrees of smoothness. \emph{Monthly 
Notices of the Royal Astronomical Society}, 492(3), p. 4019-4032.
\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{[Link]}} \cr
\item{Tibshirani (2014). Adaptive piecewise polynomial estimation via trend 
filtering. \emph{The Annals of Statistics}. 42(1), p. 285-323.
\href{https://projecteuclid.org/euclid.aos/1395234979}{[Link]}}
\item{Ramdas and Tibshirani (2016). Fast and Flexible ADMM Algorithms 
for Trend Filtering. \emph{Journal of Computational and Graphical 
Statistics}, 25(3), p. 839-858.
\href{https://amstat.tandfonline.com/doi/abs/10.1080/10618600.2015.1054033#.XfJpNpNKju0}{[Link]}}
\item{Arnold, Sadhanala, and Tibshirani (2014). Fast algorithms for 
generalized lasso problems. R package \emph{glmgen}. Version 0.0.3. 
\href{https://github.com/glmgen/glmgen}{[Link]}}
\item{Tibshirani and Taylor (2012). Degrees of freedom in lasso problems.
\emph{The Annals of Statistics}, 40(2), p. 1198-1232.
\href{https://projecteuclid.org/journals/annals-of-statistics/volume-40/issue-2/Degrees-of-freedom-in-lasso-problems/10.1214/12-AOS1003.full}{[Link]}} \cr
}
}
\seealso{
\code{\link{SURE.trendfilter}}, \code{\link{cv.trendfilter}},
\code{\link{relax.trendfilter}}, \code{\link{bootstrap.trendfilter}}
}
\author{
Collin A. Politsch, \email{collinpolitsch@gmail.com}
}
