% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendfilteringSupp-package.R
\docType{package}
\name{trendfilteringSupp-package}
\alias{trendfilteringSupp-package}
\title{Supplementary tools for data analysis with trend filtering}
\description{
This package builds on \code{glmgen} by providing additional 
             useful statistical tools for carrying out data analysis with 
             trend filtering [Tibshirani (2014)]. It contains functionality 
             for easily optimizing a trend filtering estimator by cross 
             validation or minimizing Stein's unbiased risk estimate and 
             various bootstrap algorithms for producing variability bands to 
             quantify the uncertainty in the estimator.
}
\seealso{
Refer to:
\enumerate{
\item{\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{Trend filtering – I. A modern statistical tool for time-domain astronomy and astronomical spectroscopy}}
\item{\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{Trend filtering – II. Denoising astronomical signals with varying degrees of smoothness}}
}
}
\author{
Collin A. Politsch \cr \cr 
\strong{Maintainer}: Collin A. Politsch <collinpolitsch@gmail.com>
}
\keyword{package}
