% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.trendfilter.R
\name{cv.trendfilter}
\alias{cv.trendfilter}
\title{Optimize the trend filtering hyperparameter (by V-fold cross validation)}
\usage{
cv.trendfilter(
  x,
  y,
  weights = NULL,
  k = 2L,
  nlambda = 250L,
  lambda = NULL,
  validation.error.type = c("WMAE", "WMSE", "MAE", "MSE"),
  V = 5L,
  lambda.choice = c("lambda.min", "lambda.1se"),
  n.eval = 1500L,
  x.eval = NULL,
  thinning = NULL,
  max_iter = 600L,
  obj_tol = 1e-10,
  mc.cores = detectCores()
)
}
\arguments{
\item{x}{The vector of observed values of the input variable (a.k.a. the 
predictor, covariate, explanatory variable, regressor, independent variable, 
control variable, etc.)}

\item{y}{The vector of observed values of the output variable (a.k.a. the
response, target, outcome, regressand, dependent variable, etc.).}

\item{weights}{A vector of weights for the observed outputs. These are 
defined as \code{weights = 1 / sigma^2}, where \code{sigma} is a vector of 
standard errors of the uncertainty in the output measurements. \code{weights} 
should either have length equal to 1 (corresponding to observations with a 
constant (scalar) variance of \code{sigma = 1/sqrt(weights)}) or length equal 
to \code{length(y)} (i.e. heteroskedastic outputs).}

\item{k}{The degree of the trend filtering estimator. Defaults to 
\code{k=2} (quadratic trend filtering). Must be one of \code{k = 0,1,2,3},
although \code{k=3} is discouraged due to algorithmic instability (and is
visually indistinguishable from \code{k=2} anyway).}

\item{nlambda}{The number of trend filtering hyperparameter values 
to run the grid search over. If \code{lambda = NULL}, a grid of length 
\code{nlambda} is constructed internally.}

\item{lambda}{Overrides \code{nlambda} if passed. A user-supplied vector of 
trend filtering hyperparameter values to run the grid search over. Usually
good to let them be equally-spaced in log-space. Descending order is 
encouraged to avoid ambiguity, since they will be resorted internally for
algorithmic benefits. Regardless, the sorted vector is returned in the 
function output. Using this argument is discouraged unless you know what you 
are doing.}

\item{validation.error.type}{Type of error to optimize during cross
validation. One of \code{c("WMAE","WMSE","MAE","MSE")}, i.e. mean-absolute 
deviations error, mean-squared error, and their weighted counterparts. 
If \code{weights = NULL}, then the weighted and 
unweighted counterparts are equivalent. In short, weighting helps combat
heteroskedasticity and absolute error decreases sensitivity to outliers.
Defaults to \code{"WMAE"}.}

\item{V}{The number of folds the data are split into for the V-fold cross
validation. Defaults to \code{V=5} (recommended).}

\item{lambda.choice}{One of \code{c("lambda.min","lambda.1se")}. The choice
of hyperparameter that is used for optimized trend filtering estimate. \cr \cr
\code{lambda.min}: the hyperparameter value that minimizes the cross
validation error curve. \cr \cr
\code{lambda.1se}: the largest hyperparameter value with a cross
validation error within 1 standard error of the minimum cross validation 
error. This choice therefore favors simpler (i.e. smoother) trend filtering 
estimates. The motivation here is essentially Occam's razor: the two models
yield results that are quantitatively very close, so we favor the simpler
model.}

\item{n.eval}{The length of the equally-spaced input grid to evaluate the 
optimized trend filtering estimate on.}

\item{x.eval}{Overrides \code{n.eval} if passed. A user-supplied grid of 
inputs to evaluate the optimized trend filtering estimate on.}

\item{thinning}{logical. If \code{TRUE}, then the data are preprocessed so 
that a smaller, better conditioned data set is used for fitting. When set to
\code{NULL}, the default, function will auto detect whether thinning should 
be applied (i.e., cases in which the numerical fitting algorithm will 
struggle to converge).}

\item{max_iter}{Maximum iterations allowed for the trend filtering 
convex optimization 
(\href{http://www.stat.cmu.edu/~ryantibs/papers/fasttf.pdf}{Ramdas and 
Tibshirani 2016}). Defaults to \code{max_iter = 600L}. Increase this if the 
trend filtering estimate does not appear to have fully converged to a 
reasonable estimate of the signal.}

\item{obj_tol}{The tolerance used in the convex optimization stopping 
criterion; when the relative change in the objective function is less than 
this value, the algorithm terminates. Decrease this if the trend 
filtering estimate does not appear to have fully converged to a reasonable 
estimate of the signal.}

\item{mc.cores}{Multi-core computing (for speedups): The number of cores to
utilize. Defaults to \code{min(detected.cores, V)}.}
}
\value{
An object of class 'cv.trendfilter'. This is a list with the 
following elements:
\item{x.eval}{The grid of inputs the optimized trend filtering estimate was 
evaluated on.}
\item{tf.estimate}{The optimized trend filtering estimate of the signal, 
evaluated on \code{x.eval}.}
\item{validation.method}{\code{paste0(V,"-fold CV")}}
\item{V}{The number of folds the data are split into for the V-fold cross
validation.}
\item{validation.error.type}{Type of error that validation was performed on. 
One of \code{c("WMAE","WMSE","MAE","MSE")}.}
\item{lambda}{Vector of hyperparameter values tested during validation. This
vector will always be returned in descending order, regardless of the 
ordering provided by the user. The indices \code{i.min} and \code{i.1se}
correspond to this descending ordering.}
\item{lambda.min}{Hyperparameter value that minimizes the SURE error curve.}
\item{lambda.1se}{The largest hyperparameter value that is still within one
standard error of the minimum hyperparameter's cross validation error.}
\item{lambda.choice}{One of \code{c("lambda.min","lambda.1se")}. The choice
of hyperparameter that is used for optimized trend filtering estimate.}
\item{df}{Vector of effective degrees of freedom for trend filtering
estimators fit during validation.}
\item{df.min}{The effective degrees of freedom of the optimally-tuned trend 
filtering estimator.}
\item{df.1se}{The effective degrees of freedom of the 1-stand-error rule
trend filtering estimator.}
\item{i.min}{The index of \code{lambda} that minimizes the cross validation 
error.}
\item{i.1se}{The index of \code{lambda} that gives the largest hyperparameter
value that has a cross validation error within 1 standard error of the 
minimum of the cross validation error curves.}
\item{error}{Vector of cross validation errors for the given hyperparameter 
values.}
\item{se.error}{The standard errors of the cross validation errors.
These are particularly useful for implementing the ``1-standard-error rule''. 
The 1-SE rule favors a smoother trend filtering estimate by, instead of 
using the hyperparameter that minimizes the CV error, instead uses the 
largest hyperparameter that has a CV error within 1 standard error of the
smallest CV error.}
\item{x}{The vector of the observed inputs.}
\item{y}{The vector of the observed outputs.}
\item{weights}{A vector of weights for the observed outputs. These are
defined as \code{weights = 1/sigma^2}, where \code{sigma} is a vector of 
standard errors of the uncertainty in the measured outputs.}
\item{fitted.values}{The trend filtering estimate of the signal, evaluated at
the observed inputs \code{x}.}
\item{residuals}{\code{residuals = y - fitted.values}.}
\item{k}{The degree of the trend filtering estimator.}
\item{thinning}{logical. If \code{TRUE}, then the data are preprocessed so 
that a smaller, better conditioned data set is used for fitting. When set to
\code{NULL}, the default, function will auto detect whether thinning should 
be applied (i.e., cases in which the numerical fitting algorithm will 
struggle to converge).}
\item{max_iter}{Maximum iterations allowed for the trend filtering 
convex optimization 
(\href{http://www.stat.cmu.edu/~ryantibs/papers/fasttf.pdf}{Ramdas and 
Tibshirani 2016}). Increase this if the trend filtering estimate does not 
appear to have fully converged to a reasonable estimate of the signal.}
\item{obj_tol}{The tolerance used in the convex optimization stopping 
criterion; when the relative change in the objective function is less than 
this value, the algorithm terminates. Decrease this if the trend filtering 
estimate does not appear to have fully converged to a reasonable estimate of 
the signal.}
}
\description{
\loadmathjax{} \code{cv.trendfilter} performs V-fold cross 
validation to estimate the random-input squared error of a trend filtering 
estimator over a grid of hyperparameter values and returns the full error 
curve and the optimized trend filtering estimate within a larger list of 
helpful ancillary information.
}
\details{
This will be a very detailed description... \cr \cr
\mjeqn{WMAE(\lambda) = \frac{1}{n}\sum_{i=1}^{n} |Y_i - \widehat{f}(x_i; \lambda)|\frac{\sqrt{w_i}}{\sum_j\sqrt{w_j}}}{ascii} \cr 
\mjeqn{WMSE(\lambda) = \frac{1}{n}\sum_{i=1}^{n} |Y_i - \widehat{f}(x_i; \lambda)|^2\frac{w_i}{\sum_jw_j}}{ascii} \cr 
\mjeqn{MAE(\lambda) = \frac{1}{n}\sum_{i=1}^{n} |Y_i - \widehat{f}(x_i; \lambda)|}{ascii} \cr 
\mjeqn{MSE(\lambda) = \frac{1}{n}\sum_{i=1}^{n} |Y_i - \widehat{f}(x_i; \lambda)|^2}{ascii} \cr \cr 
where \mjeqn{\widehat{f}(x_i; \lambda)}{ascii} is the trend filtering 
estimate with hyperparameter \eqn{\lambda}, evaluated at 
\mjeqn{x_i}{ascii}.
}
\examples{
#######################################################################
###  Phase-folded light curve of an eclipsing binary star system   ####
#######################################################################

# A binary star system is a pair of closely-separated stars that move
# in an orbit around a common center of mass. When a binary star system 
# is oriented in such a way that the stars eclipse one another from our 
# vantage point on Earth, we call it an 'eclipsing binary (EB) star 
# system'. From our perspective, the total brightness of an EB dips 
# periodically over time due to the stars eclipsing one another. And 
# the shape of the brightness curve is consistent within each period
# of the orbit. In order to learn about the physics of these EBs,
# astronomers 'phase-fold' the brightness curve so that all the orbital 
# periods are stacked on top of one another in a plot of the EB's phase 
# vs. its apparent brightness, and then find a 'best-fitting' model
# for the phase-folded curve. Here, we use trend filtering to fit an
# optimal phase-folded model for an EB.

data(eclipsing_binary)

# head(data)
#
# |      phase|      flux|  std.err|
# |----------:|---------:|--------:|
# | -0.4986308| 0.9384845| 0.010160|
# | -0.4978067| 0.9295757| 0.010162|
# | -0.4957892| 0.9438493| 0.010162|

cv.out <- cv.trendfilter(x = data$phase, 
                         y = data$flux, 
                         weights = 1 / data$std.err ^ 2,
                         validation.error.type = "MAE",
                         thinning = TRUE, 
                         obj_tol = 1e-14, 
                         max_iter = 1e4)

# Plot the results

par(mfrow = c(2,1), mar = c(5,4,2.5,1) + 0.1)
plot(log(cv.out$lambda), cv.out$error, main = "CV error curve", 
     xlab = "log(lambda)", ylab = "CV error")
segments(x0 = log(cv.out$lambda), x1 = log(cv.out$lambda), 
         y0 = cv.out$error - cv.out$se.error, 
         y1 = cv.out$error + cv.out$se.error)
abline(v = log(cv.out$lambda.min), lty = 2, col = "blue3")
text(x = log(cv.out$lambda.min), y = par("usr")[4], 
     labels = "optimal hyperparameter", pos = 1, col = "blue3")
plot(data$phase, data$flux, cex = 0.15, xlab = "Phase", ylab = "Flux",
     main = "Eclipsing binary phase-folded light curve")
segments(x0 = data$phase, x1 = data$phase, 
         y0 = data$flux - data$std.err, y1 = data$flux + data$std.err, 
         lwd = 0.25)
lines(cv.out$x.eval, cv.out$tf.estimate, col = "orange", lwd = 2.5)
}
\references{
\strong{Cross validation}
\enumerate{
\item Hastie, Tibshirani, and Friedman (2009). The Elements of Statistical 
Learning: Data Mining, Inference, and Prediction. 2nd edition. Springer 
Series in Statistics. 
\href{https://web.stanford.edu/~hastie/ElemStatLearn/printings/ESLII_print12_toc.pdf}{
[Online print #12]}. (See Sections 7.10 and 7.12) \cr
\item James, Witten, Hastie, and Tibshirani (2013). An Introduction to 
Statistical Learning : with Applications in R. Springer.
\href{https://www.statlearning.com/}{[Most recent online print]} (See 
Section 5.1). \emph{Less technical than the above reference.}\cr
\item Tibshirani (2013). Model selection and validation 2: Model
assessment, more cross-validation. \emph{36-462: Data Mining course notes} 
(Carnegie Mellon).
\href{https://www.stat.cmu.edu/~ryantibs/datamining/lectures/19-val2.pdf}{[Link]}
}
\strong{Trend filtering optimization algorithm}
\enumerate{
\item{Ramdas and Tibshirani (2016). Fast and Flexible ADMM Algorithms 
for Trend Filtering. \emph{Journal of Computational and Graphical 
Statistics}, 25(3), p. 839-858.
\href{https://amstat.tandfonline.com/doi/abs/10.1080/10618600.2015.1054033#.XfJpNpNKju0}{[Link]}} \cr
\item{Arnold, Sadhanala, and Tibshirani (2014). Fast algorithms for 
generalized lasso problems. R package \emph{glmgen}. Version 0.0.3. 
\href{https://github.com/glmgen/glmgen}{[Link]}} \cr
(Software implementation of Ramdas and Tibshirani algorithm) \cr
}
}
\seealso{
\code{\link{SURE.trendfilter}}, \code{\link{bootstrap.trendfilter}}
}
\author{
Collin A. Politsch, \email{collinpolitsch@gmail.com}
}
